Hide randomize groups button
It has been so long since I last worked on this that I don't remember
why the build was failing, if I even knew then. I think it could be
a very technical failure caused by the button's CSS not being accepted.
Not been an issue before but that is my thinking.

#84