Add review-groups get and edit SystemController.java methods
`editReviewGroups()` takes in a BasicDBList which is the entire review
groups collection. Each list item is a review group as each review
group was set by the client. Each review group is updated one by one
with the new values.

Of all of the methods to change and save the review groups, this one
took the most time to develop. `updateMany()` has not to my knowledge
been done in my software design class in any iteration nor Megabittron's
class, so much of this syntax came from a lot of analysis of online
documentation and examples. However, the way that some things were cast
(that was suprisingly tricky in both this controller and the request
handler) and the way exceptions are checked comes from a different file
in this repo by somebody else. I hope that the comments help. Feel free
to get a hold of me if you have questions.

Issue #100 (Implement drag and drop review group rearrangement)
Epic #84 (Review Groups)