Retrieve abstract id from list, grab abstract's details on single page
`submissionList.html` grab's a particular abstract's id, sends it
through `submissionList.component.ts` to `submissionList.service.ts`.
The service saves it as a field. `submissionView.component.ts` retrieves
that and uses it to retrieve an entire abstract's details and display
them. The `submissionList.html` just shows the title for now as a
progress placeholder.

The ID saving could be put in `submissionView.service.ts` instead, but
I think it's fine to be in the list's service for now.

Issue #132 (Navigate to individual submission from submission list)
Epic #131 (Submission Card View)