Add optional comments field to submission.ts interface
The field wasn't showing up when trying to test out where the comments
area was. The field has been added as optional as abstracts aren't
created with comments but need them once comments are added. Hoping that
the omission of the field was due to the comment interface not being
implemented yet, and not that it would complicate some server calls.
Adding a new abstract didn't mess up so it *seems* to be truly optional.
This really needs to be tested.

For now "comments?" is a string because changing the seed to a string
array broke the server, but that will eventually need to be refactored.

Issue #132 (Navigate to individual submission from submission list)
Epic #131 (Submission Card View)