Add new-comment btn and individual comment area
The server and the reviewGroups seed needs to be refactored, as comments
exist in some but not all of the seeded documents and the server
doesn't handle doing anything but showing a single comment right now.

A `"comments?": string` field was just added in the previous commit to
the submission.ts interface for abstracts. The comment fields in the
seed is just a single string, until the aforementioned server/seed
refactor happens.

Issue #132 (Navigate to individual submission from submission list)
Issue #141 (Use material colors from theme on submission page)
Epic #131 (Submission Card View)