Create expansion-panel with additional info, notes on abstract interface
Expansion panel includes all of the fields present in URSA. Many of the
fields are problematic for various reasons. TODO's annotate each one.

Some fields have different names across each document in the
seed. Ex) `submissionFormat` and `format`. Marked as `TODO: Standardize
seed`.

Some fields are alright in the seed but not codified in the ts
interface. Ex) `adviserApproval`. Marked as `TODO: Recognize field in
interface`.

Some fields are just fine, but some todos were added to make sure that
they show up as `yes/no` and not `true/false`.

`featurePresenter` in both the interface and seed should probably be
renamed. It's about if they are willing to be feature, but that field
makes it sound like they were chosen to be featured by the reviewers.

Issue #132 (Navigate to individual submission from submission list)
Epic #131 (Submission Card View)