Set gapi to null
ngOnInit() currently makes a call to the gapi object. This object isn't always fully initialized when running the tests which can lead to a test failing. This is a temporary fix. A more appropriate solution would be to have the authentication service handle the users profile image instead of trying to access it here.