Add tests for saveShirtSize(..)
This method is now tested. I'm unsure if it's properly tested, but this is how the current functionality works. This method should be changed to take a shirt size and the Id of a user. Then change the users ShirtSize to the return value of this method. Currently, the user is able to change their ShirtSize even if the http request fails and doesn't provide any feedback.