Fix mobile view issues
The sign in button should show on the initial view instead of needing to scroll down. Landscape mode on mobile caused text to overlap. I feel that this text isn't needed on mobile as most people will have used the app for the first time on desktop. The app explanation can be moved inside the app if it's needed.