Create listeners for updating user and signed in status
The loadAuth2() and initAuth2() methods don't seem to be necessary for Google sign in to work. It works with just calling renderSignIn() in LoginComponent. However, I've left them in just in case and if we wan't to update them to load and initialize the Google Client for accessing Gmail. There is also somewhat of an issue with signing out the user. When a user signs out and they click the sign in button again, it automatically signs them into the account they'd just used without providing an account picker screen. I'm not sure how much this is an issue, but may be something to keep in mind.