Move `isAdmin()` from `nav.component` to `account-info.service`
The component was checking the user's role (if they were admin) to
determine if they should see the Admin panel navigation on the nav bar.
It made more sense to make this method a part of the account info
related service.

There is a problem mocking the subscription to `authService.user$` that
is breaking the tests, but this is just a mere syntax problem that I
have to iron out.