Recognize users cursor's highlighted text
The first step to the bootleg Google Docs commenting system is
recognizing when an abstract's text is highlighted, and what text in
particular. Using the web api `Window.getSelection()`, in the way that
was suggested by Dasari Srinivas, one can highlight text and have it
show up elsewhere.

Right now, highlighted text is in a tentative spot in the comment div
as a single comment that is not stored in any way, nor even tied to the
`comment` field of `Submissions`.

Issue #147 (Create comment object)
Epic #131 (Submission Card View)