Create first draft of `comment` object
The vision of comments is that users either quote something directly
from the abstract (optional `quotedSnippet`) via highlighting a passage,
and/or making general comments (`commenterInput`). Submitters will have
commenter names that are not fellow submitters made anonymous.

Top level comments (TLC's) can be resolved or not (defaults to `false`),
and comments on TLC's may or may not need to be resolved but for now
they are optionally resolvable with the idea that they only contribute
to the TLC being resolved. Resolved comments can either be hidden, or
perhaps hidden from the normal comment space and shown in some sort of
`Resolved` area like a dropdown for posterity, or crossed out in the
normal comment sidebar? Perhaps if the TLC's exigence is argued to be
relevant again then the Resolved TLC could be toggled to be an active,
unresolved comment?

All TLC's and subcomments should be timestamped.

Issue #147 (Create comment object)
Epic #217 (Comment on Abstracts)
Comments are subset of: Epic #131 (Submission Card View)