Retrieve top lvl comments in abstracts via $oid
The topComment field in abstracts can now be retrieved from the seeds.
The `topComments.seed.json`, which is now seeded into the DB, is
referred to in the backend via the new version of `getSingleAbstract`.
The method now grabs a single abstract as an aggregate and looks up
topComments by their `$oid`. They are effectively joined together.

An example of this working is on the single submission. The toplevel
comment is now actually pulled from a seed. The subcomments are still
hardcoded.

Co-authored by: Paul Friederichsen <floogulinc@gmail.com>

Issue #147 (Create comment object)
Epic #217 (Comment on Abstracts)
Epic #131 (Submission Card View)