Retrieve topComments and attempt to refer to nested objects by ID
I can join topComments into abstracts. However, I cannot join objects
from inside of a topComment into the topComment. For example, I would
like an abstract to have an array of `topComment`, and each `topComment`
would have a `commenter` field that should refer to a `user` by the
user's `_id`.

This can be done with the commented out code in the
controller... EXCEPT the `topComment` array is no longer an array but
instead a malformed object. This is a side-effect of the `Aggregates.
unwind("$topComment")` that is needed to be able to join individual
`user` ids to a topComment's `commenter` field. If the `topComment`
field could be reverted to an array again, this wouldn't be a problem.

Issue #147 (Create comment object)
Epic #217 (Comment on Abstracts)
Epic #131 (Submission Card View)