Nest `topComments` in abstracts
Almost all of a topComment is now nested in an abstract, using IDs to
join other objects. The last piece is getting users to be able to joined
inside of the `subComments` field `commenter`.

The `group` method needs to re-add all fields via `first`, but for now
it's easier to debug without all of those there.

Issue #147 (Create comment object)
Epic #217 (Comment on Abstracts)
Epic #131 (Submission Card View)