Update user roles successfully
Rather big commit, but the major issue solved in this commit led to
refactoring almost the whole stack touching related to user object
updating.

The way things work now as intended: user object is grabbed manually
when selecting a role from the "Change role" dropdown on a particular
user. The `Role` field is set to be updated in the frontend and then
updated in the backend.

The issue for the longest time that I had was that we stopped
serializing json using JSON.serialize() months ago in favor of what was
considered to be at the time a more up to date way of doing things.
It "worked" for the longest time because we weren't doing things with
retrieving users related to `_ids` until now. Perhaps there is a way to
match our new method but I don't have the time to figure that out.
Instead, I brought in the serializeIterable provided to us in 3601 and
it just works. I'm not going to pretend to understand the difference
between JSON.serialize() and serializeIterable().

Git blame if you are curious: https://github.com/Megabittron/DeploymentExperiment-URS/commit/e51fd1c80ac77ac1cd6f629c20b6a01ac8ad6c3b#diff-89391bfcd07f18dd4fabdf21049a53e4L71 (edited)

Issue #225 (Edit user roles)
Epic #10 (Account Information)