Fix editAbstract checkboxes bug
The checkboxes would initialize with their given value, but would
not update the visual state (checked or not) until a second click.
However, the state actually changed with each click. Therefore, the
visual checkbox and actual state were the exact opposite of what the
user intended to check.

A fix to this is provided by the below link... down to changing
`(click)` to `(change)` and explicitly checking `[checked]` values
to be `true`.

Fix courtesy of Hrishikesh Kale,
April 10 '18 @https://stackoverflow.com/a/49749782

Issue #262 (Edit abstracts as student user)
Epic #131 (Submission Card View)