Add home page details (pres types/categories/timeline)
The timeline has strikethroughs dependent on dates passed. The exact
hours are up in the air. The way that the dates are checked are
admittedly not so great, but the "proper" way of doing if/else
using `*ngIf` and `ng-template` wasn't handling `li`'s well.
The field names for the various date conditions are also quite verbose,
but mostly for emphasis on how out of date the code will be next year.

Issue #268 (handle student user permissions separate of admins)
Epic #10 (Account Information)