Clarify role of viewer of individual abstract
BUG: Cannot navigate to list of abstracts that a user should be able to
access if they are a chair or reviewer, let alone to a single
abstract view. Users and admins are fine. Likely a silly bug in the
frontend due to the quick development in this branch related to this
user permissions issue.

Going to bed now... two energy drinks and two coffees this evening
was both the best and worst idea...

Issue #268 (handle student user permissions separate of admins)
Epic #10 (Account Information)