Allow 10 copresenters
This is a really clutch commit of HUGE changes
but I have little time. Here goes:

Uses dynamic forms to make 10 copresenters
possible. Email validated. The submitter is included
in the FormArray of presenters by default, so the
indexes that prevent more than 10 more copresenters
also hides that the submitter was added as a
copresenter. Busted, I know, I know. But this is
otherwise a major improvement on how flexible the
code in general regarding things like this is.

This also updates how certain fields show up on
the submission review and individual submission view.
Also, the review submission step text is flush left
in a centered element because the <ul> of presenters
was ugly when centered.