Rehaul submission form around new object structure, clean review step
All the hardcoded advisor/presenter fields are now unified into
an object for a presenter and an object for an advisor. As such,
html reflects that by iterating over how many things are in each
array.

Was having a difficult time figuring out how to make a [stepControl]
be dependent on multiple form groups. While it was a bit of a copout,
moving the presenters fields out of the category/feature step fixed
any roadblocks. Besides, it makes sense to be broken out anyways.

The wrapup step is formatted in a nicer fashion, in particular with
arrays of data. Presenters and advisors are now in a bulleted list,
fields with one display value each like sponsoredOrganizations are
now properly comma seperated.