Stylize and add various logic to new submissions
Added a lot of logic that either prevents entering empty
forms, in particular the disciplines. Trims out whitespace
in various forms like that to prevent junk data being
sent or accidentally shown on the end page. These need
to be refined, but I added an edge case for the most important
one being the disciplines that makes going back to that
page and doing it right be what is needed.

Renamed a lot of the form controls to ones that make
more sense.

What used to be many text areas are now proper inputs.
Added mat-labels and some placeholders.Made the info page look
very nice and put my email on it. Bolded all the review prompts.